requirements.txt
# MetaApi Dependencies
# Core Flask framework
Flask==2.3.3
Werkzeug==2.3.7

# MetaTrader 5 integration
MetaTrader5==5.0.45

# Data processing
pandas==2.1.3
numpy==1.24.4

# Telegram bot
pyTelegramBotAPI==4.14.0

# Logging with colors
colorlog==6.8.0

# HTTP requests (if needed for external APIs)
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# Type hints support for older Python versions
typing-extensions==4.8.0

# Development dependencies (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.8
